fix 쿠키 만료 시간을 변수로 전달

+++ src/main/java/com/themoment/officialgsm/global/util/CookieUtil.java
@@ -9,9 +9,9 @@ import org.springframework.stereotype.Component;
 @Slf4j
 @Component
 public class CookieUtil {
-    public static void addRefreshTokenCookie(HttpServletResponse response, String name, String value, Integer maxAge, boolean httpOnly){
+    public static void addRefreshTokenCookie(HttpServletResponse response, String name, String value, Long maxAge, boolean httpOnly){
         Cookie cookie = new Cookie(name, value);
-        cookie.setMaxAge(maxAge);
+        cookie.setMaxAge(Math.toIntExact(maxAge));
         cookie.setHttpOnly(httpOnly);
         response.addCookie(cookie);
     }
